moved ConfigManager
